+++ src/includes/plugin.php
@@ -6,9 +6,9 @@
  * @copyright WebSharks™
  */
 declare (strict_types = 1);
-namespace WebSharks\WpSharks\WooCommerce\s2MemberX\Pro;
+namespace WebSharks\WpSharks\WooCommerce\Restrictions\Pro;
 
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes\App;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes\App;
 
 if (!defined('WPINC')) {
     exit('Do NOT access this file directly: '.basename(__FILE__));
@@ -20,6 +20,6 @@ add_action('plugins_loaded', function () {
         require_once __DIR__.'/stub.php';
         new App(); // Plugin instance.
     } else {
-        wp_sharks_core_rv_notice('WooCommerce Restrictions — s2Member X');
+        wp_sharks_core_rv_notice('WooCommerce Restrictions Pro');
     }
 });