+++ src/includes/templates/s-core/notices/on-install.php
@@ -6,15 +6,15 @@
  * @copyright WebSharks™
  */
 declare (strict_types = 1);
-namespace WebSharks\WpSharks\WooCommerce\s2MemberX\Pro;
+namespace WebSharks\WpSharks\WooCommerce\Restrictions\Pro;
 
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes;
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Interfaces;
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Traits;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Interfaces;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Traits;
 #
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes\AppFacades as a;
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes\SCoreFacades as s;
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes\CoreFacades as c;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes\AppFacades as a;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes\SCoreFacades as s;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes\CoreFacades as c;
 #
 use WebSharks\WpSharks\Core\Classes as SCoreClasses;
 use WebSharks\WpSharks\Core\Interfaces as SCoreInterfaces;
@@ -31,6 +31,6 @@ use function get_defined_vars as vars;
 extract($this->vars); // Template variables.
 ?>
 <p>
-    <?= sprintf(__('<strong>%1$s</strong> v%2$s installed successfully.', 'woocommerce-s2member-x'), esc_html($this->App->Config->©brand['©name']), esc_html($this->App::VERSION)) ?><br />
-    <?= sprintf(__('~ Start by protecting some of your content: <a href="%1$s" class="button" style="text-decoration:none;">Create Restriction</a>', 'woocommerce-s2member-x'), esc_url(a::createRestrictionUrl())) ?>
+    <?= sprintf(__('<strong>%1$s</strong> v%2$s installed successfully.', 'woocommerce-restrictions'), esc_html($this->App->Config->©brand['©name']), esc_html($this->App::VERSION)) ?><br />
+    <?= sprintf(__('~ Start by protecting some of your content: <a href="%1$s" class="button" style="text-decoration:none;">Create Restriction</a>', 'woocommerce-restrictions'), esc_url(a::createRestrictionUrl())) ?>
 </p>