+++ src/includes/uninstall.php
@@ -6,9 +6,9 @@
  * @copyright WebSharks™
  */
 declare (strict_types = 1);
-namespace WebSharks\WpSharks\WooCommerce\s2MemberX\Pro;
+namespace WebSharks\WpSharks\WooCommerce\Restrictions\Pro;
 
-use WebSharks\WpSharks\WooCommerce\s2MemberX\Pro\Classes\App;
+use WebSharks\WpSharks\WooCommerce\Restrictions\Pro\Classes\App;
 
 if (!defined('WPINC')) {
     exit('Do NOT access this file directly: '.basename(__FILE__));