+++ tests/includes/phpunit.php
@@ -6,7 +6,7 @@
  * @copyright WebSharks™
  */
 declare (strict_types = 1);
-namespace WebSharks\WpSharks\WooCommerce\s2MemberX\Pro;
+namespace WebSharks\WpSharks\WooCommerce\Restrictions\Pro;
 
 require_once $_SERVER['CI_NFO_PHPUNIT_BOOTSTRAP'];
 require_once $_SERVER['CI_NFO_WWW_DIR'].'/wp-load.php';